Feat: 헬로박스 회원정보이름 받아오기 완료
