+++ src/components/mypage/HelloBox.tsx
@@ -1,16 +1,24 @@
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
-import { useAppSelector } from '../../app/hooks';
-import { userDataEmail } from '../../features/userData/userDataSlice';
+import { useAppSelector, useAppDispatch } from '../../app/hooks';
+import { userDataName } from '../../features/userData/userDataSlice';
 
 const HelloBox = () => {
-  const userEmail = useAppSelector(userDataEmail);
-  console.log('유저이메일:' + userEmail);
+  const userName = useAppSelector(userDataName);
+
   return (
     <UserImg>
-      {/* <img src={datas.userInfo.profileImage} alt="" style={{ width: 100 }} /> */}
+      {/* <img
+        src={
+          datas.userInfo.profileImage.includes('null')
+            ? process.env.PUBLIC_URL + '/icons/ic-member.svg'
+            : datas.userInfo.profileImage
+        }
+        alt=""
+        style={{ width: 100 }}
+      /> */}
       <div>
-        {/* <h3>{datas.userInfo.name} 님 안녕하세요 ˙ᵕ˙</h3> */}
+        <h3>{userName} 님 안녕하세요 ˙ᵕ˙</h3>
         <p>누적 결제금액 : 원</p>
       </div>
     </UserImg>
